+++ matmul_vector_1d.cpp
@@ -12,8 +12,8 @@
 #include <vector>
 
 // Naive row-major matrix multiplication: C = A (m x k) * B (k x n).
-std::vector<double> matmul_naive(const std::vector<double>& A,
-                                 const std::vector<double>& B,
+std::vector<float> matmul_naive(const std::vector<float>& A,
+                                const std::vector<float>& B,
                                  size_t m,
                                  size_t k,
                                  size_t n) {
@@ -24,12 +24,12 @@ std::vector<double> matmul_naive(const std::vector<double>& A,
         throw std::invalid_argument("B size does not match k*n");
     }
 
-    std::vector<double> C(m * n, 0.0);
+    std::vector<float> C(m * n, 0.0f);
 
     // Naive i-j-k triple loop.
     for (size_t i = 0; i < m; ++i) {
         for (size_t j = 0; j < n; ++j) {
-            double sum = 0.0;
+            float sum = 0.0f;
             for (size_t p = 0; p < k; ++p) {
                 sum += A[i * k + p] * B[p * n + j];
             }
@@ -63,18 +63,18 @@ size_t safe_mul(size_t a, size_t b, const char* label) {
     return a * b;
 }
 
-double calculate_gflops(size_t m,
-                        size_t n,
-                        size_t k,
-                        std::chrono::duration<double, std::milli> elapsed_ms) {
-    const double elapsed_seconds = elapsed_ms.count() / 1000.0;
-    if (elapsed_seconds <= 0.0) {
-        return 0.0;
+float calculate_gflops(size_t m,
+                       size_t n,
+                       size_t k,
+                       std::chrono::duration<float, std::milli> elapsed_ms) {
+    const float elapsed_seconds = elapsed_ms.count() / 1000.0f;
+    if (elapsed_seconds <= 0.0f) {
+        return 0.0f;
     }
 
-    const double flops = 2.0 * static_cast<double>(m) * static_cast<double>(n) *
-                         static_cast<double>(k);
-    return flops / elapsed_seconds / 1e9;
+    const float flops = 2.0f * static_cast<float>(m) * static_cast<float>(n) *
+                        static_cast<float>(k);
+    return flops / elapsed_seconds / 1e9f;
 }
 
 int main(int argc, char** argv) {
@@ -116,13 +116,13 @@ int main(int argc, char** argv) {
     const size_t b_size = safe_mul(k, n, "B size");
     const size_t c_size = safe_mul(m, n, "C size");
 
-    std::vector<double> A(a_size);
-    std::vector<double> B(b_size);
-    std::vector<double> C;
+    std::vector<float> A(a_size);
+    std::vector<float> B(b_size);
+    std::vector<float> C;
 
     // Deterministic random initialization.
     std::mt19937 rng(12345);
-    std::uniform_real_distribution<double> dist(0.0, 1.0);
+    std::uniform_real_distribution<float> dist(0.0f, 1.0f);
     for (size_t i = 0; i < a_size; ++i) {
         A[i] = dist(rng);
     }
@@ -135,11 +135,11 @@ int main(int argc, char** argv) {
     C = matmul_naive(A, B, m, k, n);
 
     const auto end = std::chrono::steady_clock::now();
-    const std::chrono::duration<double, std::milli> elapsed_ms = end - start;
-    const double gflops = calculate_gflops(m, n, k, elapsed_ms);
+    const std::chrono::duration<float, std::milli> elapsed_ms = end - start;
+    const float gflops = calculate_gflops(m, n, k, elapsed_ms);
 
     // Small checksum to keep output compact for large matrices.
-    double checksum = 0.0;
+    float checksum = 0.0f;
     for (size_t i = 0; i < c_size; ++i) {
         checksum += C[i];
     }
